requirements.txt
langchain_core
langchain_community
python_dotenv
streamlit
langserve
fastapi
uvicorn
pypdf
faiss-cpu
langchain-groq
unstructured
unstructured[md]
libmagic
langchain_ollama
langchain_experimental
tabulate
einops
sentence-transformers
